Oooohkay... Lots of changes here. See desc.
Key exchange is working. First message sent by client is login name. Second
message is public key. Server then checks to see if key is in username's
authorized_keys file. If so, opens /etc/chatd/keys/id_rsa.pub and sends it. No
encryption yet, will be easy to add (it's one function call) later on and the
system is designed to be able to handle it being added.